The ubsan preset had not compiled since C1, so nothing had run it
`cmake --build --preset microbrowser-ubsan` failed on four `-Wsign-conversion`
errors in two UTF-8 decoders (`EncodingBindings.cpp`, `ConstructableStylesheets
.cpp`, both from `c52dfbf`): `(lead & 0x0F) << 12 | …` promotes to `int` and
assigns to a `std::uint32_t`. The perf and debug presets do not carry `-Werror`,
so the tree built everywhere anyone looked and the sanitizer that catches
shift-overflow and misaligned reads had been silently unrunnable.

Fixed by casting each piece before the shift rather than the result after it,
which is also what makes the shift itself well-defined. It is unrelated to the
URL work in the commits around it and is its own commit for that reason -- but
it is *why* those commits could not be checked under UBSan until now.

  ubsan: 2097/2097, zero runtime errors
  asan:  2097/2097, zero memory errors (the 368 bytes of leak are
         ConstructableStylesheets' shared_ptr slot and predate this branch)

+++ src/bindings/ConstructableStylesheets.cpp
@@ -274,13 +274,15 @@ void DomBindings::InstallCssOm() {
         cp = (lead & 0x1F) << 6 | (static_cast<unsigned char>(input[i + 1]) & 0x3F);
         width = 2;
       } else if ((lead & 0xF0) == 0xE0 && i + 2 < input.size()) {
-        cp = (lead & 0x0F) << 12 | (static_cast<unsigned char>(input[i + 1]) & 0x3F) << 6 |
-             (static_cast<unsigned char>(input[i + 2]) & 0x3F);
+        cp = static_cast<std::uint32_t>(lead & 0x0F) << 12 |
+             static_cast<std::uint32_t>(static_cast<unsigned char>(input[i + 1]) & 0x3F) << 6 |
+             static_cast<std::uint32_t>(static_cast<unsigned char>(input[i + 2]) & 0x3F);
         width = 3;
       } else if ((lead & 0xF8) == 0xF0 && i + 3 < input.size()) {
-        cp = (lead & 0x07) << 18 | (static_cast<unsigned char>(input[i + 1]) & 0x3F) << 12 |
-             (static_cast<unsigned char>(input[i + 2]) & 0x3F) << 6 |
-             (static_cast<unsigned char>(input[i + 3]) & 0x3F);
+        cp = static_cast<std::uint32_t>(lead & 0x07) << 18 |
+             static_cast<std::uint32_t>(static_cast<unsigned char>(input[i + 1]) & 0x3F) << 12 |
+             static_cast<std::uint32_t>(static_cast<unsigned char>(input[i + 2]) & 0x3F) << 6 |
+             static_cast<std::uint32_t>(static_cast<unsigned char>(input[i + 3]) & 0x3F);
         width = 4;
       } else {
         char buf[8];
